Update UseCases.md
Made changes to employee use case descriptions.

For editing/deleting games I changed the flows so that the user would view the list of games and click the button for the action they wished to perform on a particular game, instead of having multiple pages for each action which all essentially just list all of the games. Editing/deleting events was given the same treatment. This was mostly just a matter of changing the wording, the flows themselves are still more or less the same so all related diagrams should be okay but it may be worth comparing to ensure consistency.